requirements.txt
asgiref==3.4.1
coverage==6.2
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
